Drop unused hook file, drop localrc plans (for now)
`localrc` and some related environment things do not make sense as they
stand, and have massively complex implementation strategies.

I think it makes sense to migrate to a self-contained command that
spawns a sub-shell rather than trying to fuck around with all the
contrived state management bullshit I would have to do to make it all
work.

Spawning a subshell does come with a few advantages. Just off the top of
my head:

* The direnv-file at work consumes VAULT_ADDR, but also sets it to a
  default value if it isn't defined. This leads to bullshit where
  ```
  cd ..
  export VAULT_ADDR=<redacted>
  cd repo
  ```
  is required, because if you just
  ```
  export VAULT_ADDR=<redacted>
  direnv reload
  ```
  the VAULT_ADDR is unset first, because it's exported by the
  direnv-file. This isn't great, but it is a feature of direnv, and one
  that was annoying enough that it spawned v1 of this proposal
  (localrc/dirrc, I didn't bother reading to see what I settled on).

  Any `localrc`-like strategy will fall victim to this as well, and
  there's no real way around it.

  However, if a subshell is spawned, we get around this. If
  `.untitled_rcfile` contains
  ```
  export VAULT_ADDR=whatever
  ```

  then it'll be set by the time the shell spawns, and direnv kicks in.
  This also allows for the environment management system I imagined in
  umbra v1 (which is different from localrc/dirrc v1, which is part of
  umbra v2)
* No need to care about unloading, because everything will be contained
  in the subshell. Want to unload? Close the subshell, and you'll be
  dropped back in the parent shell

  This is why it enables teh environment management system.
* Cuts down on raw shell scripts the user has to deal with, and avoids
  needing to wrap umbra in a shell script or some contrived bullshit
  like that. Environment stuff can't propagate from C++ to the outer
  shell without a `source`d shell between umbra and the user, and that
  just sucks.

And one advantage that's also a disadvantage: it'll likely require
manual user input to start. The advantage is that it's easier to keep
the security bits in check here. If the files have changed, ask
interactively before picking which files to load.

The disadvantage is the exact same thing; it being manual means it's
easier to forget. However, this could still work with the native shell
chdir hooks. Going for `eval $(umbra load-hooks)` in the .<shell>rc
would  sitll work, but it isn't required for it to work, so it'll
avoid concerns around use of `eval`.

On the third hand, it might make sense to not have `load-hooks` at all,
but require the user to set it up themselves with templates in a doc
file.
I don't know which I prefer. I think I might push that down the road
until I get a feel for manual usage.

========================================================================
Versioning tangent

* umbra v1 refers to v1 of the umbra design spec, which was discarded
  and mostly deleted. I have a copy laying around for however long I
  keep the SSD in this computer
* umbra v2 refers to v2 of the umbra design spec, which is the current
  one.

v2 will be the final version, because it actually works as a spec. This
doesn't mean umbra will never change, it just means umbra's core design
system works, and can be further iterated. There may be what would
technically be a v3 at some point in the future, but because that'll be
reached iteratively rather than through a full destroying the entire
spec and starting from scratch, it doesn't count in the system I use
here.

Similarly, localrc v1 is a now discarded spec. Unlike umbra, it won't be
back as a v2, because the alternate design is another module outright.